updates due to new Fallen-8 version

+++ Fallen-8 Intro/Config.cs
@@ -11,18 +11,18 @@ namespace Intro
 
         #region Node properties
 
-        public const Int32 W_ID_PROPERTY_ID = 0;
-        public const Int32 WORD_PROPERTY_ID = 1;
+        public const UInt16 W_ID_PROPERTY_ID = 0;
+        public const UInt16 WORD_PROPERTY_ID = 1;
 
-        public const Int32 CO_N_EDGE_PROPERTY_ID = 23;
-        public const Int32 CO_S_EDGE_PROPERTY_ID = 42;
+        public const UInt16 CO_N_EDGE_PROPERTY_ID = 23;
+        public const UInt16 CO_S_EDGE_PROPERTY_ID = 42;
 
         #endregion
 
         #region Edge properties
 
-        public const Int32 SIG_PROPERTY_ID = 10;
-        public const Int32 FREQ_PROPERTY_ID = 11;
+        public const UInt16 SIG_PROPERTY_ID = 10;
+        public const UInt16 FREQ_PROPERTY_ID = 11;
 
         #endregion
     }
